requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
python-jose==3.3.0
pydantic[email]==2.6.3
python-multipart==0.0.9
alembic==1.13.1
pandas==2.2.1
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
